Running this bare spec takes 0.2 seconds.
We should be able to cut a massive amount of time
off this if we avoid hitting the database.

Let's start with the index action.

If you tail the test log, you'll see that this is performing
an insert into the database on each test run, which is then
rolled back.

We can easily stub out this behaviour.

In older versions of RSpec we would have written:

  Widget.stub(all: widgets)

But the more modern syntax:

  allow(Widget).to receive(:all).and_return(widgets)
or

  expect(Widget).to receive(:all).and_return(widgets)

The difference between these is that the latter
will raise a failed expectation error, while the former
will not complain.

There are different situations where each is appropriate.

Here I've chosen to use 'expect' as I believe it will result
in a more helpful message if the spec fails.

Re-running this spec with the test log open will demonstrate
that it is no longer hitting the database.