Let's move onto the `show` action.
This expects a widget to exist in the database.

We'll use a double again, named widget_double to avoid
confusion.

The origin generator spec called #to_param to obtain
and ID to use, but we can just pick an arbitrary number.

We set up the expectation, and we have to quote the ID
since it's passed as a String from the controller to the model's
`find` method.