The `post` action is little tricker. I think this is because Rails violates "tell, don't ask" here - it tells Widget to instantiate a new instance, then asks it if it can be saved (i.e. if it is valid).
I'm not going to try re-architect Rails here, so let's work within
those limitations.

We'll say that we expect the newly instantiated widget to receive
a `save` message. We don't specify the return value, so by default
it's nil, which means the controller will treat this is an invalid
object. But we will cover that behaviour in other specs.