Again we are checking the type of assigned Widget which seems unnecessary. We'll remove this spec.
We can no longer use the before block so we'll
add the `post` action back into each spec
within this context.

Note that we are now using mock_model instead of
a plain double. This is necessary because
Rails calls `model_name` on the double at some
point, and trying to stub out this would be messy.

Note also that we specify the 'Widget', the parameter
for mock_model, as a String. This means we could write
the controller spec without yet even having any model,
which can be useful for outside-in development.